Worked on networking for PC
- Kind of confused on how responses work for PC. I need to figure that
out for status and ping
- Created new packages in the request and response packages to seperate
MCPE protocols and MCPC protocols
- Got a new kitten today :D
