+++ src/main/java/net/redstonelamp/request/pc/StatusRequest.java
@@ -1,0 +1,16 @@
+package net.redstonelamp.request.pc;
+
+import net.redstonelamp.request.Request;
+
+public class StatusRequest extends Request {
+	
+	public int protocol;
+	public String address;
+	public int port;
+
+	@Override
+	public void execute() {
+		//TODO?
+	}
+	
+}
