+++ src/main/java/net/redstonelamp/response/pc/StatusResponse.java
@@ -1,0 +1,5 @@
+package net.redstonelamp.response.pc;
+
+public class StatusResponse {
+
+}
