+++ src/main/java/net/redstonelamp/response/pe/v27/ChatResponse.java
@@ -14,7 +14,9 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with RedstoneLamp.  If not, see <http://www.gnu.org/licenses/>.
  */
-package net.redstonelamp.response;
+package net.redstonelamp.response.pe.v27;
+
+import net.redstonelamp.response.Response;
 
 public class ChatResponse extends Response{
